Restructurize folders
