+++ src/fractal/mod.rs
@@ -1,0 +1,3 @@
+pub mod haar_tree;
+pub mod image;
+pub mod variants;
