+++ src/utils/coordinate.rs
@@ -6,10 +6,10 @@ pub struct Coord {
     pub y: i32,
 }
 
-impl ops::Add<Coord> for Coord {
-    type Output = Coord;
-    fn add(self, rhs: Coord) -> Self::Output {
-        Coord {
+impl ops::Add<Self> for Coord {
+    type Output = Self;
+    fn add(self, rhs: Self) -> Self::Output {
+        Self {
             x: self.x + rhs.x,
             y: self.y + rhs.y,
         }
