+++ src/utils/mod.rs
@@ -1,2 +1,2 @@
-pub mod ans;
 pub mod bitwise;
+pub mod coordinate;
